+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -58,6 +58,8 @@ exclude: true
   ([bug 1621094]({{bugzilla}}1621094))
 - Added [`ContentBlockingController`][70.24], accessible via [`GeckoRuntime.getContentBlockingController`][70.25]
   to allow modification and inspection of a content blocking exception list.
+- Added support for exposing the content blocking log in [`ContentBlockingController`][70.26].
+  ([bug 1580201]({{bugzilla}}1580201))
 
 [70.1]: {{javadoc_uri}}/GeckoSessionSettings.Builder.html#contextId-java.lang.String-
 [70.2]: {{javadoc_uri}}/StorageController.html#clearDataForSessionContext-java.lang.String-
@@ -84,6 +86,7 @@ exclude: true
 [70.23]: {{javadoc_uri}}/GeckoSession.ContentDelegate.html
 [70.24]: {{javadoc_uri}}/ContentBlockingController.html
 [70.25]: {{javadoc_uri}}/GeckoRuntime.html#getContentBlockingController--
+[70.26]: {{javadoc_uri}}/ContentBlockingController.html
 
 ## v69
 - Modified behavior of ['setAutomaticFontSizeAdjustment'][69.1] so that it no 
@@ -327,4 +330,4 @@ exclude: true
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: f2d89facd57470b760be9a867173d02bf9040c75
+[api-version]: c51e338dbc7220c0fe2676826b36e1bb346d58bc